Added new groups to break up views in a more organized way
