+++ Shared/Views/Onboarding/OnboardingView.swift
@@ -49,29 +49,6 @@ struct OnboardingView: View {
             })
         }
     }
-    
-    struct OnboardingDetailRow: View {
-        let symbol: String
-        let text: String
-        
-        var body: some View {
-            HStack {
-                Image(systemName: symbol)
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: 32)
-                    .scaledToFit()
-                    .padding()
-                    .foregroundColor(.mobYellow)
-                
-                Text(text)
-                    .foregroundColor(.mobGreen)
-                    .font(.body)
-                
-            }
-        }
-    }
-
 }
 
 struct OnboardingView_Previews: PreviewProvider {
